+++ packages/class-core/src/deltas/types.g.ts
@@ -20,6 +20,7 @@
 
 import {
     Containment,
+    Node,
     Property,
     Reference,
     SingleRef
@@ -254,7 +255,7 @@ export class ReferenceAddedDelta implements IDelta {
         public readonly parent: INodeBase,
         public readonly reference: Reference,
         public readonly index: number,
-        public readonly newReference: SingleRef<INodeBase>
+        public readonly newReference: SingleRef<Node>
     ) {
     }
 }
@@ -264,7 +265,7 @@ export class ReferenceDeletedDelta implements IDelta {
         public readonly parent: INodeBase,
         public readonly reference: Reference,
         public readonly index: number,
-        public readonly deletedReference: SingleRef<INodeBase>
+        public readonly deletedReference: SingleRef<Node>
     ) {
     }
 }
@@ -274,8 +275,8 @@ export class ReferenceChangedDelta implements IDelta {
         public readonly parent: INodeBase,
         public readonly reference: Reference,
         public readonly index: number,
-        public readonly newReference: SingleRef<INodeBase>,
-        public readonly oldReference: SingleRef<INodeBase>
+        public readonly newReference: SingleRef<Node>,
+        public readonly oldReference: SingleRef<Node>
     ) {
     }
 }