+++ packages/core/src/reading.ts
@@ -17,12 +17,14 @@ export type ResolveInfoDeducer<NT extends Node> = (node: NT, reference: Referenc
 
 /**
  * An interface that's used to parametrize generic serialization of
- * (in-memory) nodes of the given type (parameter).
+ * (in-memory) nodes of the given type (parameter) `NT`.
  * Implementations of these interfaces {w|c}ould be:
  *  - specific to LionCore (so to match m3/types.ts)
  *  - generic to serialize {@link DynamicNode dynamic nodes}
+ * The `RT` type (parameter) is the type of any *referenced* node,
+ * which could e.g. be wider – most likely – or narrower than `NT`.
  */
-export interface Reader<NT extends Node> {
+export interface Reader<NT extends Node, RT extends Node = NT> {
 
     /**
      * @return The {@link Concept concept} of the given node
@@ -45,7 +47,7 @@ export interface Reader<NT extends Node> {
      * @return The string value of the `resolveInfo` field of a {@link LionWebJsonReferenceTarget serialized reference target},
      * or `undefined` to indicate that no `resolveInfo` could be derived.
      */
-    resolveInfoFor?: ResolveInfoDeducer<NT>
+    resolveInfoFor?: ResolveInfoDeducer<RT>
 
 }
 
@@ -54,5 +56,5 @@ export interface Reader<NT extends Node> {
  *
  * @deprecated Use {@link Reader} instead.
  */
-export interface ExtractionFacade<NT extends Node> extends Reader<NT> {}
+export interface ExtractionFacade<NT extends Node, RT extends Node> extends Reader<NT, RT> {}